requirements.txt
streamlit
pdfplumber
python-dotenv
langchain-google-genai